awss3sink: added UploaderPartCache history with tests
The UploaderPartCache stores data about the buffers
and sizes that have been sent to AWS over the mutli-
part uploader.  The design retaints a copy of the
buffers that fall within the limits of 'depth' as well
as the size of the buffer (whether retained or not)
so that various features can be leveraged to support
a partial seek functionality.

Signed-off-by: Thomas Goodwin <thomas.goodwin@laerdal.com>