awss3sink: implemented seeking query
As noted in comments, the reply here is not soomething
we can accurately implement because there may be times
when the seekable region is actually in two different
ranges: the current buffer and the buffers within the
cache.  Splitting the difference then, the reply will
be "everything" if the cache is configured to retain
any buffers, and "nothing" / not seekable if is not
(the default).

Signed-off-by: Thomas Goodwin <thomas.goodwin@laerdal.com>